[client] fix rebase errors
