+++ client/bot-community.py
@@ -6,8 +6,8 @@ from random_words import RandomWords
 from math import floor
 
 from py_client.client import Client
-from py_client.client.ipfs import Ipfs
-from py_client.communities import populate_locations, generate_community_spec
+from py_client.ipfs import Ipfs
+from py_client.communities import populate_locations, generate_community_spec, meta_json
 
 NUMBER_OF_LOCATIONS = 100
 MAX_POPULATION = 12 * NUMBER_OF_LOCATIONS
