+++ client/test-locations-mediterranean.json
@@ -4,7 +4,7 @@
     "meta": {
       "name": "Mediterranea",
       "symbol": "MTA",
-      "icons": "41eSfKJrhrR6CYxPfUbwAN18R77WbxXoViRWQMAF4hJB"
+      "icons": "QmVmew4gZHyCK2Fv4UBgsvfLdf1Q6UiF9MD6wsfPCuNVQp"
     },
     "bootstrappers": [
       "5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY",
